requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
fastapi
    # via mcp-project (pyproject.toml)
    # via pydantic
    # via pydantic-settings
    # via uvicorn
fastmcp==2.5.1
annotated-types==0.7.0
    # via pydantic
anthropic==0.52.0
    # via mcp-project (pyproject.toml)
anyio==4.9.0
    # via
    #   anthropic
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
arxiv==2.2.0
    # via mcp-project (pyproject.toml)
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
distro==1.9.0
    # via anthropic
feedparser==6.0.11
    # via arxiv
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   mcp
httpx-sse==0.4.0
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.10.0
    # via anthropic
mcp==1.9.2.dev7+f2f4dbd
    # via mcp-project (pyproject.toml)
nest-asyncio==1.6.0
    # via mcp-project (pyproject.toml)
pydantic==2.11.5
    # via
    #   anthropic
    #   mcp
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via mcp
python-dotenv==1.1.0
    # via
    #   mcp-project (pyproject.toml)
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
requests==2.32.3
    # via arxiv
sgmllib3k==1.0.0
    # via feedparser
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
sse-starlette==2.3.5
    # via mcp
starlette==0.46.2
    # via
    #   mcp
    #   sse-starlette
typing-extensions==4.13.2
    # via
    #   anthropic
    #   anyio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via mcp
